LLE: Warning cleanup, config fixes, and dead code removal (-1,322 lines)
- Clean build with zero warnings on macOS Clang
- Fix config dedup_strategy value mismatch (underscore vs hyphen)
- Remove stale prompt.style from default config template
- Remove dead readline syntax highlighting code:
  - Phase 3 performance optimization structs
  - Smart triggering system
  - Highlighting buffer management
  - Word boundary highlighting functions
  - Unused public API stubs
- Add MAYBE_UNUSED macros for spec-compliant unused functions
- All 51 tests pass

+++ src/lle/core/performance.c
@@ -39,7 +39,9 @@
  * Note: lle_get_thread_id is defined in error_handling.c, we use a local helper
  */
 static uint32_t perf_get_thread_id(void) {
-    return (uint32_t)pthread_self();
+    /* Cast through uintptr_t first for portability - pthread_t may be a
+     * pointer type on some platforms (e.g., macOS) */
+    return (uint32_t)(uintptr_t)pthread_self();
 }
 
 /**
